workflow: make compiler warnings hard Fixes #601

+++ include/molsys/simulationBox.hpp
@@ -83,7 +83,7 @@ namespace molsys
 
         linalg::Vec3D                      _centerOfMass = {0.0, 0.0, 0.0};
         std::vector<std::shared_ptr<Atom>> _atoms;
-        std::vector<int>                   _innerRegionCenterAtomIndices;
+        std::vector<size_t>                _innerRegionCenterAtomIndices;
         std::vector<Molecule>              _molecules;
         std::vector<MoleculeType>          _moleculeTypes;
 
@@ -169,7 +169,7 @@ namespace molsys
          * QMMM related methods *
          ************************/
 
-        void addInnerRegionCenterAtoms(const std::vector<int>& atomIndices);
+        void addInnerRegionCenterAtoms(const std::vector<size_t>& atomIndices);
         void setupForcedCoreMolecules(const std::vector<int>& moleculeIndices);
         void setupForcedLayerMolecules(const std::vector<int>& moleculeIndices);
         void setupForcedOuterMolecules(const std::vector<int>& moleculeIndices);
@@ -196,7 +196,7 @@ namespace molsys
         [[nodiscard]] double                 getTotalCharge() const;
         [[nodiscard]] double                 getDensity() const;
         [[nodiscard]] linalg::Vec3D&         getCenterOfMass();
-        [[nodiscard]] std::vector<int>       getInnerRegionCenterAtomIndices();
+        [[nodiscard]] std::vector<size_t>    getInnerRegionCenterAtomIndices();
 
         [[nodiscard]] Atom&         getAtom(size_t index);
         [[nodiscard]] Molecule&     getMolecule(size_t index);
